requirements.txt
# Core FastAPI Application
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.30.0
alembic>=1.12.0

# Authentication and Security
PyJWT>=2.8.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Configuration and Environment
python-dotenv>=1.0.0

# HTTP Client and Async Support
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0

# Data Processing and Analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# WebSocket Support
websockets>=11.0.0

# CORS and Multipart Support
python-multipart>=0.0.6

# Caching
redis>=5.0.0

# AWS SDK
boto3>=1.34.0
aioboto3>=12.0.0
botocore>=1.34.0

# Agent Framework Dependencies
langchain>=0.1.0
langchain-aws>=0.1.0
langgraph>=0.2.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langsmith>=0.1.0

# Amazon Bedrock AgentCore Runtime
bedrock-agentcore>=1.0.0
strands-agents>=1.0.0
bedrock-agentcore-starter-toolkit>=1.0.0

# Retry Logic and Utilities
tenacity>=8.2.0

# Logging and Monitoring
structlog>=23.1.0

# Date and Time Utilities
python-dateutil>=2.8.0

# JSON and Data Serialization
orjson>=3.9.0

# Financial Data
yfinance>=0.2.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Visualization
matplotlib>=3.8.0
plotly>=5.17.0

# Additional AWS and Cloud Dependencies
aws-cdk-lib>=2.100.0
constructs>=10.3.0

# Additional utilities that might be needed
click>=8.1.0
rich>=13.0.0
typer>=0.9.0

# Testing dependencies
moto>=4.2.0
docker>=6.1.0
pytest-bdd>=7.0.0